Combined socket control with speech_interactor'
